requirements.txt
tensorflow==2.18.0
streamlit==1.44.1
numpy==2.0.2
pillow==11.1.0